scripts/requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
idna==3.10
libclang==18.1.1
loguru==0.7.2
pycryptodomex==3.22.0
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.17.0
Unidecode==1.3.8
urllib3==2.2.3
win32-setctime==1.1.0
